Found the converging issue, that is from the initial peak finding
The n=9 peak finding in K_Mean failed and it falls back to resutl of n=3, which induces this weird feedback behaviour since the grid is all messed up

Fix the issue found above by
Increasing the number of time the k-means algorithm will be run with different centroid seeds, which is n_init parameter in K-Means function. This helps the algorithm to iterate through more possible output and it is thus more likely to hit the correct on.

Increase this parameter from 10 to 25 makes the clustering of 9x9 grid work. Previously the n=9 clustering failed and it falls back to n=3, which is the pattern we saw above

Commit:
git@github.com:UCBoulder/ChimeraGenTools.git
7e8b928e7f4d6b616c8a1d3d1c00f3fd21cca4c6